3단계 - 문자열 계산기
