+++ domain/src/main/java/com/camp/nextstep/edu/domain/TokenType.kt
@@ -1,0 +1,8 @@
+package com.camp.nextstep.edu.domain
+
+sealed class TokenType(val operator: String)  {
+    object Plus: TokenType("+")
+    object Minus: TokenType("-")
+    object Multiply: TokenType("*")
+    object Divide: TokenType("/")
+}
